nEW Changes

+++ FinalProject/business logic/CarsLogic.cs
@@ -12,6 +12,7 @@ namespace Business_Logic
     /// </summary>
     class CarsLogic : BaseLogic
     {
+        #region AllAboutCars
 
         public List<Car> GetAllCars()
         {
@@ -19,7 +20,9 @@ namespace Business_Logic
             {
                 //TODO - remember to add a "where" IsDeleted = false, this will be used for "delete" cars.
                 //return DB.CarFleets.Select(car => car.CarID).ToArray();
+
                 var query = DB.SelectAllCarsIsNotDeleted();
+
                 List<Car> carList = null;
 
                 foreach (var item in query)
@@ -30,6 +33,7 @@ namespace Business_Logic
                     car.Photo = item.Photo;
                     car.Transmission = item.Transmission;
                     car.ManufacturerName = item.ManufacturerName;
+                    car.ReadyForRental = item.ReadyForRental;
                     carList.Add(car);
                 }
                 return carList;
@@ -43,16 +47,20 @@ namespace Business_Logic
             }
         }
 
-        public void AddCar( string CarID, int CarModelID, int Mileage, string Photo, bool ReadyForRental)
+        public bool AddNewCar(string CarID, int CarModelID, int Mileage, string Photo, bool ReadyForRental)
         {
+            var checkIfExist = DB.CarFleets.Where(p => p.CarID == CarID).FirstOrDefault();
             try
             {
-                DB.AddCarToCarFleet(CarID, CarModelID, Mileage, Photo, ReadyForRental, false);
-
+                if (checkIfExist == null)
+                {
+                    DB.AddCarToCarFleet(CarID, CarModelID, Mileage, Photo, ReadyForRental, false);
+                    return true;
+                }
+                return false;
             }
             catch (Exception)
             {
-
                 throw;
             }
             
@@ -76,25 +84,59 @@ namespace Business_Logic
 
         public void DeleteCar(string CarID)
         {
+            var query = DB.CarFleets.Where(s => s.CarID == CarID).FirstOrDefault();
+
             try
             {
+                if (query != null)
                 DB.DeleteCarFormFleet(CarID);
             }
             catch (Exception)
             {
-
                 throw;
             }
 
 
         }
 
-        public string[] GetAllCarManufacture()
+        public bool ReturnCarAndChangeToReadyForRent(string CarID)
         {
+            var query = DB.CarFleets.Where(s => s.CarID == CarID).FirstOrDefault();
+
+            if (query != null)
+            {
+                query.ReadyForRental = true;
+
+                DB.SaveChanges();
+
+                return true;
+            }
+            return false;
+        }
+
+
+
+        #endregion
+
+        #region AllAboutManufacturers
+
+        public List<Manufacturer> GetAllCarManufacture()
+        {
+            var query = DB.SelectAllManufacturers();
+
+            List<Manufacturer> manufacturerList = null;
             try
             {
-               return DB.Manufactures.Select(M => M.ManufacturerName).ToArray();
+                foreach (var item in query)
+                {
+                    Manufacturer manufacturer = new Manufacturer();
+                    manufacturer.ManufacturerID = item.ManufacturerID;
+                    manufacturer.ManufacturerName = item.ManufacturerName;
+                    manufacturerList.Add(manufacturer);
+                }
+                return manufacturerList;
             }
+
             catch (Exception)
             {
 
@@ -102,12 +144,18 @@ namespace Business_Logic
             }
         }
 
-        public void AddManufacturer(string manufactureName)
+        public bool AddManufacturer(string manufactureName)
         {
             try
             {
-                DB.AddManufacturer(manufactureName);
+                var checkIfExist = DB.Manufactures.Where(m => m.ManufacturerName == manufactureName).FirstOrDefault();
 
+                if (checkIfExist == null)
+                {
+                    DB.AddManufacturer(manufactureName);
+                    return true;
+                }
+                return false;
             }
             catch (Exception)
             {
@@ -120,7 +168,12 @@ namespace Business_Logic
         {
             try
             {
-                DB.UpdateManufacturer(manufactureID, manufactureName);
+                var checkIExist = DB.Manufactures.Where(m => m.ManufacturerName == manufactureName).FirstOrDefault();
+
+                if (checkIExist == null)
+                {
+                    DB.UpdateManufacturer(manufactureID, manufactureName);
+                }
             }
             catch (Exception)
             {
@@ -129,6 +182,38 @@ namespace Business_Logic
             }
         }
 
+        public void DeleteManufacturer(int manufactureID)
+        {
+            var query = DB.Manufactures.Where(s => s.ManufacturerID == manufactureID).FirstOrDefault();
+
+            if (query != null)
+            {
+                query.IsDelete = true;
+
+                var carsModelsQuery = DB.CarsModels.Where(p => p.ManufacturerID == query.ManufacturerID).ToList();
+
+                if (carsModelsQuery != null)
+                {
+                    foreach (var item in carsModelsQuery)
+                    {
+                        item.IsDelete = true;
+
+                        var carQuery = DB.CarFleets.Where(x => x.CarModelID == item.CarModelID).FirstOrDefault();
+
+                        if (carQuery != null)
+                        {
+                            carQuery.ReadyForRental = false;
+                            carQuery.IsDelete = true;
+                        }
+                    }
+                }
+
+                DB.SaveChanges();
+            }
+        }
+        #endregion
+
+        #region CarModels
         public int[] GetAllCarModels()
         {
             try
@@ -152,7 +237,7 @@ namespace Business_Logic
             DB.UpdateCarModel(carModelID,manufacturerID, carModel, manufactureYear, transmission, dailyRate, lateFee);
         }
 
-        
+        #endregion
